# net/http: move mux.mu.RLock() and mux.mu.RUnlock to shouldRedirectRLocked method