--- a/src/net/http/server.go
+++ b/src/net/http/server.go
@@ -2391,9 +2391,7 @@ func (mux *ServeMux) match(path string) (h Handler, pattern string) {
 // not for path itself. If the path needs appending to, it creates a new
 // URL, setting the path to u.Path + "/" and returning true to indicate so.
 func (mux *ServeMux) redirectToPathSlash(host, path string, u *url.URL) (*url.URL, bool) {
-	mux.mu.RLock()
 	shouldRedirect := mux.shouldRedirectRLocked(host, path)
-	mux.mu.RUnlock()
 	if !shouldRedirect {
 		return u, false
 	}
@@ -2406,6 +2404,8 @@ func (mux *ServeMux) redirectToPathSlash(host, path string, u *url.URL) (*url.UR
 // path+"/". This should happen if a handler is registered for path+"/" but
 // not path -- see comments at ServeMux.
 func (mux *ServeMux) shouldRedirectRLocked(host, path string) bool {
+	mux.mu.RLock()
+	defer mux.mu.RUnlock()
 	p := []string{path, host + path}
 
 	for _, c := range p {