--- a/src/net/http/server.go
+++ b/src/net/http/server.go
@@ -2391,7 +2391,7 @@ func (mux *ServeMux) match(path string) (h Handler, pattern string) {
 // not for path itself. If the path needs appending to, it creates a new
 // URL, setting the path to u.Path + "/" and returning true to indicate so.
 func (mux *ServeMux) redirectToPathSlash(host, path string, u *url.URL) (*url.URL, bool) {
-	shouldRedirect := mux.shouldRedirectRLocked(host, path)
+	shouldRedirect := mux.shouldRedirect(host, path)
 	if !shouldRedirect {
 		return u, false
 	}
@@ -2400,14 +2400,14 @@ func (mux *ServeMux) redirectToPathSlash(host, path string, u *url.URL) (*url.UR
 	return u, true
 }
 
-// shouldRedirectRLocked reports whether the given path and host should be redirected to
+// shouldRedirect reports whether the given path and host should be redirected to
 // path+"/". This should happen if a handler is registered for path+"/" but
 // not path -- see comments at ServeMux.
-func (mux *ServeMux) shouldRedirectRLocked(host, path string) bool {
+func (mux *ServeMux) shouldRedirect(host, path string) bool {
 	mux.mu.RLock()
 	defer mux.mu.RUnlock()
 	p := []string{path, host + path}
-
+	
 	for _, c := range p {
 		if _, exist := mux.m[c]; exist {
 			return false