Update shouldRedirectRLocked method

`RLock` and `Runlock` move to `shouldRedirectRLocked`
renamed `shouldRedirectRLocked` to `shouldRedirect`


Benchmark:

BenchmarkServeMux-8
old    26017 ns/op	   17280 B/op	     360 allocs/op
new    25420 ns/op	   17280 B/op	     360 allocs/op

BenchmarkServeMux_SkipServe-8
old 14834 ns/op
new 15386 ns/op